Break LeftPanel into smaller components
