+++ react-redux-app/src/App.js
@@ -1,10 +1,13 @@
-import React from "react";
-import { createStore, applyMiddleware } from "redux";
-import { Provider } from "react-redux";
-import thunk from "redux-thunk";
+import React, { useEffect } from "react";
+// import { createStore, applyMiddleware } from "redux";
+// import { Provider } from "react-redux";
+// import thunk from "redux-thunk";
 import styled from "styled-components";
+import { connect } from "react-redux";
+import { useGetPokemon } from "./hooks/useGetPokemon";
 
-import { pokemonReducer as reducer } from "./reducers";
+import { getPokemon } from "./actions";
+// import { pokemonReducer as reducer } from "./reducers";
 
 // import PokemonList from "./components/PokemonList";
 // import GenButtons from "./components/GenButtons";
@@ -23,27 +26,53 @@ const AppContainer = styled.div`
     display: flex;
 `;
 
-const store = createStore(reducer, applyMiddleware(thunk));
+// const store = createStore(reducer, applyMiddleware(thunk));
 
 // click green "Submit" in RightPanel to trigger API call to fetch pokemon info based on the number in NumInput in RightPanel
 // set res.data.name to NameScreen in LeftPanel
 // set res.data.order to DexNum in LeftPanel
 // set
 
-function App() {
+const App = props => {
+    const {
+        getInputNum,
+        // setInputNum: setNum,
+        // decrementInputNum,
+        // incrementInputNum,
+        getEndpoint
+    } = useGetPokemon();
+
+    const handleEnterKeyGetPokemon = e => {
+        if (e.key === "Enter") {
+            console.log("getInputNum in App: ", getInputNum());
+            console.log("getEndpoint: ", getEndpoint());
+            props.getPokemon(getEndpoint());
+        }
+    };
+
+    useEffect(() => {
+        window.addEventListener("keypress", handleEnterKeyGetPokemon);
+    }, [() => getEndpoint()]);
+
     return (
-        <Provider store={store}>
-            <AppContainer>
-                {/* <h1>Basic Pokedex</h1>
+        // <Provider store={store}>
+        <AppContainer>
+            {/* <h1>Basic Pokedex</h1>
                 <GenButtons />
                 <PokemonList /> */}
 
-                <LeftPanel />
-                <Divider />
-                <RightPanel />
-            </AppContainer>
-        </Provider>
+            <LeftPanel />
+            <Divider />
+            <RightPanel />
+        </AppContainer>
+        //  </Provider>
     );
-}
+};
 
-export default App;
+const mapStateToProps = state => {
+    return {
+        pokemon: state.pokemon
+    };
+};
+
+export default connect(mapStateToProps, { getPokemon })(App);
